Extracts base class for tenant exports ⚾️
OX-10040
